cording_question/requirements.txt
click==7.1.2
pdf2image==1.14.0
Pillow==8.1.2
pyocr==0.8
